+++ eventservice/api/utils/group_auth.py
@@ -4,6 +4,7 @@ import logging
 from flask import g
 
 logger = logging.getLogger(__name__)
+ALL_GROUP_EVENTS = 'all_group-events'
 
 
 def generate_groups_request():
@@ -81,3 +82,13 @@ def check_permission_access_event(event, include_private_events, group_ids):
         if event and event.get('isGroupPrivate') is True:
             return False
     return True
+
+
+def check_all_group_event_admin():
+    is_all_group_event = False
+    if 'user_token' in g and not g.user_token_data.get('anonymous'):
+        if g.user_token_data.get('permissions'):
+            if g.user_token_data.get('permissions').lower().find(ALL_GROUP_EVENTS) != -1:
+                is_all_group_event = True
+
+    return is_all_group_event